TextFormat: support both FMP7 FA/F formats and FMP4 format.
context: https://github.com/Wohlstand/OPN2BankEditor/pull/99

Upon further investigation, it turned out the existing format is based on
FMP7 syntax, which is technically incompatible with FMP4.
@jpcima mentioned that there are two formats for PMD, so this change
follows that approach too.

The relevant format documentation can be found (in Japanese) at:

- http://fmpdoc.fmp.jp/fm%e9%9f%b3%e6%ba%90%e9%9f%b3%e8%89%b2%e5%ae%9a%e7%be%a9opna/
- http://fmpdoc.fmp.jp/fm%e9%9f%b3%e6%ba%90%e9%9f%b3%e8%89%b2%e5%ae%9a%e7%be%a9opm/
- http://fmpdoc.fmp.jp/fm%e9%9f%b3%e6%ba%90%e9%9f%b3%e8%89%b2%e5%ae%9a%e7%be%a9/

Now you can see there is another "OPM compatible" format (which is NOT
the same format as the existing code had implemented).
Ideally it had better be supported too, but that needs "dt2" parameter
implementation in the synthesizer, which is not likely compatible with
the *actual* OPM emulator.

OPM support in FMP7 is an "extended" emulation (http://guu.fmp.jp/archives/143),
so any unexpected audio output would be anticipated anyways.
It is therefore skipped in this change so far.